Move to opaque llvm pointers
For a long time llvm has been slowly moving to a model where pointers
are an opaque type, instead of knowing the pointee type. This has been
complete for a while now, so they've started deprecating the APIs for
constructing a pointer with a pointee type (these APIs ignore the
pointee type anyway), or doing pointer casts, which are no-ops. This PR
removes use of those APIs and is a nice simplification of some code.

+++ src/Func.cpp
@@ -907,7 +907,7 @@ Func Stage::rfactor(const vector<pair<RVar, Var>> &preserved) {
 
     ReductionDomain rdom{definition.schedule().rvars(), definition.predicate(), true};
     SubstitutionMap rdom_promises;
-    for (int i = 0; i < rdom.domain().size(); i++) {
+    for (int i = 0; i < (int)rdom.domain().size(); i++) {
         const auto &[var, min, extent] = rdom.domain()[i];
         rdom_promises.emplace(var, promise_clamped(RVar(rdom, i), min, min + extent - 1));
     }
